work with MovieList

+++ app/src/main/java/com/twins/testapp/model/Movie.kt
@@ -11,6 +11,6 @@ data class Movie(
     val poster_path: String?,
     val release_date: String?,
     val title: String?,
-    val vote_average: Int?,
+    val vote_average: Double?,
     val video: Boolean?,
 )
